refactor(global): ThirdPartyScope를 OAuthScope로 통합
