--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/application/repository/ThirdPartyScopeJpaRepository.kt
@@ -1,13 +1,0 @@
-package team.themoment.datagsm.common.domain.application.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import team.themoment.datagsm.common.domain.application.entity.ThirdPartyScopeJpaEntity
-
-interface ThirdPartyScopeJpaRepository : JpaRepository<ThirdPartyScopeJpaEntity, Long> {
-    fun findByApplicationIdAndScopeName(
-        applicationId: String,
-        scopeName: String,
-    ): ThirdPartyScopeJpaEntity?
-
-    fun findAllByApplicationIdIn(applicationIds: Set<String>): List<ThirdPartyScopeJpaEntity>
-}
