+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/client/dto/response/OAuthScopeListResDto.kt
@@ -1,0 +1,5 @@
+package team.themoment.datagsm.common.domain.client.dto.response
+
+data class OAuthScopeListResDto(
+    val list: List<OAuthScopeResDto>,
+)
