+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/client/entity/constant/OAuthScope.kt
@@ -2,15 +2,14 @@ package team.themoment.datagsm.common.domain.client.entity.constant
 
 import org.springframework.security.core.GrantedAuthority
 
-open class OAuthScope(
-    open val scope: String,
-    open val description: String,
+class OAuthScope(
+    val applicationId: String,
+    val scopeName: String,
+    val description: String,
 ) : GrantedAuthority {
+    val scope: String get() = "$applicationId:$scopeName"
+
     override fun getAuthority(): String = "SCOPE_$scope"
-
-    val category: String get() = scope.substringBefore(':')
-
-    val categoryDisplayName: String get() = CATEGORY_DISPLAY_NAMES[category] ?: category
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -21,14 +20,14 @@ open class OAuthScope(
     override fun hashCode(): Int = scope.hashCode()
 
     companion object {
-        val SELF_READ = OAuthScope("self:read", "내 정보 조회")
-        val builtinValues: List<OAuthScope> = listOf(SELF_READ)
-        private val CATEGORY_DISPLAY_NAMES = mapOf("self" to "사용자")
-
-        fun fromString(scope: String): OAuthScope? = builtinValues.find { it.scope == scope }
-
-        fun getAllScopes(): Set<String> = builtinValues.map { it.scope }.toSet()
-
-        fun groupByCategory(scopes: List<OAuthScope>): Map<String, List<OAuthScope>> = scopes.groupBy { it.categoryDisplayName }
+        fun fromScopeString(scopeStr: String): OAuthScope? {
+            val colonIdx = scopeStr.indexOf(':')
+            if (colonIdx <= 0 || colonIdx == scopeStr.lastIndex) return null
+            return OAuthScope(
+                applicationId = scopeStr.substring(0, colonIdx),
+                scopeName = scopeStr.substring(colonIdx + 1),
+                description = "",
+            )
+        }
     }
 }
